Sessions and Transactions: use Java-8 SAM interfaces

+++ Common/src/intermediate/AdaptIntermediateSession.java
@@ -58,7 +58,7 @@ public class AdaptIntermediateSession implements IntegralIntermediateSession {
 
 
   @Override
-  public long ping() {
+  public int ping() {
     if (myRemoteSession.isClosed())
       throw new DBSessionIsClosedException("The intermediate session is closed.");
