+++ DekafMain/src/core/DBLeasedSession.kt
@@ -9,28 +9,22 @@ interface DBLeasedSession : DBSession, DBTransactionControl {
 
     /**
      * Checks whether the session is really alive.
-
-     *
      *
      * If the check is successful, just returns the ping duration.
      *
-     *
-     *
      * If the check failed, closes the session and throws an exception.
      *
-
      * @return ping duration (in milliseconds).
-     * *
-     * *
+     *
      * @see .isClosed
      */
-    fun ping(): Long
+    fun ping(): Int
 
     /**
      * Check whether the session is closed.
      * @return  is closed.
-     * *
-     * *
+     *
+     *
      * @see .ping
      * @see .close
      */
@@ -38,12 +32,9 @@ interface DBLeasedSession : DBSession, DBTransactionControl {
 
     /**
      * Closes the session.
-
-     *
      *
      * When session is already closed, just does nothing.
      *
-
      * @see .isClosed
      */
     fun close()
